fix(datagrid): move key nav utils file to utils folder
