+++ projects/angular/src/data/datagrid/utils/key-navigation-strategies/expanded-row.ts
@@ -6,8 +6,8 @@
  */
 
 import { CellCoordinates } from '../key-navigation-grid.controller';
+import { KeyNavigationUtils } from '../key-navigation-utils';
 import { DefaultKeyNavigationStrategy } from './default';
-import { KeyNavigationUtils } from './key-nav-utils';
 
 export class ExpandedRowKeyNavigationStrategy extends DefaultKeyNavigationStrategy {
   constructor(utils: KeyNavigationUtils) {
